Reorganize utils folder with wallet and parity subfolders
