+++ lib/assets/transactions/approve.js
@@ -5,7 +5,7 @@ import trace from "../../utils/generic/trace";
 import findEventInLog from "../../utils/ethereum/findEventInLog";
 import getTokenContract from "../contracts/getTokenContract";
 import toProcessable from "../utils/toProcessable";
-import sendTransaction from "../../utils/sendTransaction";
+import sendTransaction from "../../utils/parity/sendTransaction";
 
 import type { Address } from "../schemas/Address";
 import type { TokenSymbol } from "../schemas/TokenSymbol";
