+++ lib/assets/transactions/transferTo.js
@@ -4,7 +4,7 @@ import findEventInLog from "../../utils/ethereum/findEventInLog";
 import getTokenContract from "../contracts/getTokenContract";
 import toProcessable from "../utils/toProcessable";
 import setup from "../../utils/setup";
-import sendTransaction from "../../utils/sendTransaction";
+import sendTransaction from "../../utils/parity/sendTransaction";
 
 import type { Address } from "../schemas/Address";
 import type { TokenSymbol } from "../schemas/TokenSymbol";
