+++ lib/exchange/transactions/cancelOrder.js
@@ -2,7 +2,7 @@
 import setup from "../../utils/setup";
 import getSimpleMarketContract from "../contracts/getSimpleMarketContract";
 import findEventInLog from "../../utils/ethereum/findEventInLog";
-import sendTransaction from "../../utils/sendTransaction";
+import sendTransaction from "../../utils/parity/sendTransaction";
 
 import type { Address } from "../../assets/schemas/Address";
