+++ lib/exchange/transactions/makeOrder.js
@@ -7,7 +7,7 @@ import getOrder from "../../exchange/calls/getOrder";
 import toProcessable from "../../assets/utils/toProcessable";
 import getAddress from "../../assets/utils/getAddress";
 import approve from "../../assets/transactions/approve";
-import sendTransaction from "../../utils/sendTransaction";
+import sendTransaction from "../../utils/parity/sendTransaction";
 
 import type { Address } from "../../assets/schemas/Address";
 import type { Order } from "../schemas/Order";
