+++ lib/fund/transactions/convertUnclaimedRewards.js
@@ -5,7 +5,7 @@ import setup from "../../utils/setup";
 import getFundContract from "../contracts/getFundContract";
 import ensure from "../../utils/generic/ensure";
 import findEventInLog from "../../utils/ethereum/findEventInLog";
-import sendTransaction from "../../utils/sendTransaction";
+import sendTransaction from "../../utils/parity/sendTransaction";
 
 import type { Address } from "../../assets/schemas/Address";
