+++ lib/fund/transactions/makeOrderFromFund.js
@@ -7,7 +7,7 @@ import toProcessable from "../../assets/utils/toProcessable";
 import getAddress from "../../assets/utils/getAddress";
 import findEventInLog from "../../utils/ethereum/findEventInLog";
 import getFundContract from "../contracts/getFundContract";
-import sendTransaction from "../../utils/sendTransaction";
+import sendTransaction from "../../utils/parity/sendTransaction";
 
 import type { Address } from "../../assets/schemas/Address";
 import type { TokenSymbol } from "../../assets/schemas/TokenSymbol";
