+++ lib/fund/transactions/shutDownFund.js
@@ -2,7 +2,7 @@
 import setup from "../../utils/setup";
 import getFundContract from "../contracts/getFundContract";
 import ensure from "../../utils/generic/ensure";
-import sendTransaction from "../../utils/sendTransaction";
+import sendTransaction from "../../utils/parity/sendTransaction";
 
 import type { Address } from "../../assets/schemas/Address";
