+++ lib/fund/transactions/takeOrderFromFund.js
@@ -4,7 +4,7 @@ import BigNumber from "bignumber.js";
 import setup from "../../utils/setup";
 import toProcessable from "../../assets/utils/toProcessable";
 import getOrder from "../../exchange/calls/getOrder";
-import sendTransaction from "../../utils/sendTransaction";
+import sendTransaction from "../../utils/parity/sendTransaction";
 import findEventInLog from "../../utils/ethereum/findEventInLog";
 import ensure from "../../utils/generic/ensure";
 import getFundContract from "../contracts/getFundContract";
