+++ lib/fund/transactions/toggleSubscription.js
@@ -1,6 +1,6 @@
 // @flow
 import setup from "../../utils/setup";
-import sendTransaction from "../../utils/sendTransaction";
+import sendTransaction from "../../utils/parity/sendTransaction";
 import getFundContract from "../contracts/getFundContract";
 import ensure from "../../utils/generic/ensure";
