+++ lib/participation/transactions/executeRequest.js
@@ -4,7 +4,7 @@ import BigNumber from "bignumber.js";
 import setup from "../../utils/setup";
 import findEventInLog from "../../utils/ethereum/findEventInLog";
 import toReadable from "../../assets/utils/toReadable";
-import sendTransaction from "../../utils/sendTransaction";
+import sendTransaction from "../../utils/parity/sendTransaction";
 import getFundContract from "../../fund/contracts/getFundContract";
 
 import type { Address } from "../../assets/schemas/Address";
