+++ lib/participation/transactions/list.js
@@ -1,7 +1,7 @@
 // @flow
 import ensure from "../../utils/generic/ensure";
 import getParticipationContract from "../contracts/getParticipationContract";
-import sendTransaction from "../../utils/sendTransaction";
+import sendTransaction from "../../utils/parity/sendTransaction";
 
 import type { Address } from "../../assets/schemas/Address";
