+++ lib/participation/transactions/redeem.js
@@ -8,7 +8,7 @@ import findEventInLog from "../../utils/ethereum/findEventInLog";
 import ensure from "../../utils/generic/ensure";
 import approve from "../../assets/transactions/approve";
 import getFundContract from "../../fund/contracts/getFundContract";
-import sendTransaction from "../../utils/sendTransaction";
+import sendTransaction from "../../utils/parity/sendTransaction";
 
 import type { Address } from "../../assets/schemas/Address";
