+++ lib/participation/transactions/subscribe.js
@@ -3,7 +3,7 @@ import BigNumber from "bignumber.js";
 
 import setup from "../../utils/setup";
 import ensure from "../../utils/generic/ensure";
-import sendTransaction from "../../utils/sendTransaction";
+import sendTransaction from "../../utils/parity/sendTransaction";
 import findEventInLog from "../../utils/ethereum/findEventInLog";
 import approve from "../../assets/transactions/approve";
 import getBalance from "../../assets/calls/getBalance";
