--- lib/utils/importWallet.js
@@ -1,5 +1,0 @@
-import Wallet from "ethers-wallet";
-
-const importWallet = mnemonic => Wallet.Wallet.fromMnemonic(mnemonic);
-
-export default importWallet;
