+++ lib/version/transactions/setupFund.js
@@ -1,7 +1,7 @@
 // @flow
 import addressBook from "@melonproject/protocol/address-book.json";
 import setup from "../../utils/setup";
-import sendTransaction from "../../utils/sendTransaction";
+import sendTransaction from "../../utils/parity/sendTransaction";
 import getVersionContract from "../contracts/getVersionContract";
 import getFundInformations from "../../fund/calls/getFundInformations";
 import findEventInLog from "../../utils/ethereum/findEventInLog";
